discharge_slip/requirements.txt
asgiref==3.2.10
Django==3.1
html5lib==1.1
numpy==1.19.1
Pillow==7.2.0
PyPDF2==1.26.0
python-dateutil==2.8.1
python3-ghostscript==0.5.0
pytz==2020.1
pywin32==228
reportlab==3.5.48
six==1.15.0
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.4
